chore(oauth2): move quickstart files into dedicated folder
