+++ .coveragerc
@@ -13,5 +13,4 @@ exclude_lines =
 # Files to exclude from consideration
 omit =
     # Utilities for OAuth2 connectors testing
-    oauth_connector_quickstart.py
-    toucan_connectors/oauth2_connector/oauth2_authorization_webserver.py
+    toucan_connectors/oauth2_connector/quickstart
